Zadanie 2/requirements.txt
uvicorn
fastapi
psycopg2-binary
tzdata
typing
pydantic-settings
